+++ TS/Ticket.cpp
@@ -5,15 +5,15 @@
 #include "Ticket.h"
 
 Ticket::Ticket() {
-    memset(key, 0x00, CryptoPP::AES::DEFAULT_KEYLENGTH);
-    memset(iv, 0x00, CryptoPP::AES::BLOCKSIZE);
+    memset(key, 0x00, (size_t) CryptoPP::AES::DEFAULT_KEYLENGTH);
+    memset(iv, 0x00, (size_t) CryptoPP::AES::BLOCKSIZE);
 }
 
 std::string Ticket::createTicket(std::string privilegeInfo) {
     std::string plaintext = privilegeInfo;
     std::string ciphertext;
 
-    CryptoPP::AES::Encryption aesEncryption(key, CryptoPP::AES::DEFAULT_KEYLENGTH);
+    CryptoPP::AES::Encryption aesEncryption(key, (size_t) CryptoPP::AES::DEFAULT_KEYLENGTH);
     CryptoPP::CBC_Mode_ExternalCipher::Encryption cbcEncryption(aesEncryption, iv);
 
     CryptoPP::StreamTransformationFilter stfEncryptor(cbcEncryption, new CryptoPP::StringSink(ciphertext));
